requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
Naked==0.1.31
pycryptodome==3.14.1
pyfiglet==0.8.post1
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
urllib3==1.26.9
